type all server side code
